requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
biopython==1.70
click==6.7
cython==0.27.1
decorator==4.1.2          # via networkx
joblib==0.11              # via pomegranate
networkx==1.9
numpy==1.13.3
pandas==0.20.3
pomegranate==0.7.7
python-dateutil==2.6.1    # via pandas
pytz==2017.2              # via pandas
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0               # via python-dateutil
